Update goto anything-related shortcuts for Sublime Text
The current description of Command + P for Sublime Text just
indicates that it can be used to open files. However, it is much
more powerful than that. You can use it to navigate to a specific
line in a different file, or even see all methods available in whatever
file is present in your file structure.

This link has a good overview of "Goto Anything" along with associated
operators:

http://sublime-text-unofficial-documentation.readthedocs.org/en/latest/file_management/file_management.html